Create Finite Map data structure
The trick is the same as with finite sets. In fact, finite sets are
now be implemented in terms of finite maps!

Also do some general refactoring - the `Ord` typeclass is now in it's
own module. Also identified duplication in my original proof that
insertion commutes and restructured to remove it (main idea is that we
can assume WLOG that k1 < k2 in a lemma and then dispatch to that lemma
appropriately after actually comparing k1 and k2 - i.e: with k1 and k2
perhaps swapped)